pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "KernelAgent"
version = "0.1.0"
description = "Autonomous Triton Kernel Generation Agent"
readme = "README.md"
requires-python = ">=3.8,<3.13"
license = {text = "MIT"}
authors = [
    {name = "PyTorch Labs"},
]
keywords = ["triton", "gpu", "kernel", "generation", "llm", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai",
    "jinja2",
    "python-dotenv",
    "gradio",
    "pathlib2>=2.3.0; python_version < '3.4'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
]

[project.scripts]
kernel-agent = "triton_ui:main"

[project.urls]
"Homepage" = "https://github.com/pytorch-labs/KernelAgent"
"Bug Tracker" = "https://github.com/pytorch-labs/KernelAgent/issues"
"Repository" = "https://github.com/pytorch-labs/KernelAgent"

[tool.setuptools.packages.find]
include = ["triton_kernel_agent*", "kernel_perf_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=triton_kernel_agent --cov-report=term-missing"
